pyproject.toml
[project]
name = "wind-farm-assistant"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
dependencies = [
    "mcp",
    "boto3",
    "pandas",
    "numpy",
    "scipy",
    "requests",
    "pillow",
    "strands-agents==1.15.0",
    "strands-agents-tools",
    "seaborn",
    "termcolor",
    "turbine-models",
    "folium",
    "shapely",
    "geopandas",
    "py-wake",
    "ipykernel>=6.30.1",
    "networkx>=3.4.2",
    "folium>=0.20.0",
    "shapely>=2.1.1",
    "bedrock-agentcore>=0.1.2",
    "bedrock-agentcore-starter-toolkit>=0.1.6",
    "pillow",
    "termcolor>=3.1.0",
    "matplotlib",
    "weasyprint",
    "markdown",
    "jupyter>=1.1.1",
    "PyPDF2",  # if you're working with PDFs
    "typing",
    "branca>=0.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
